[PR](collision system): Fix ennemies damage collision

+++ src/server/network/ServerNetworkSystem.hpp
@@ -292,7 +292,7 @@ class ServerNetworkSystem {
                                   network::DisconnectReason reason);
     void processExpiredGracePeriods();
     void finalizeDisconnection(std::uint32_t userId);
-    void handleClientInput(std::uint32_t userId, std::uint8_t inputMask);
+    void handleClientInput(std::uint32_t userId, std::uint16_t inputMask);
     void handleClientChat(std::uint32_t userId, const std::string& message);
     void handleGetUsersRequest(std::uint32_t userId);
